Split out patch vertex
